% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.flys_sections}
\alias{df.flys_sections}
\title{Reference gauging stations according to FLYS3}
\format{
A \code{data.frame} with 24 rows and 4 variables:
\describe{
  \item{river}{name of the FLYS3 water body (type \code{character}).}
  \item{gauging_station}{name of the reference gauging station (type \code{character}).}
  \item{from}{uppermost station of the river section (type \code{numeric}).}
  \item{to}{lowermost station of the river section (type \code{numeric}).}
  \item{uuid}{name of the reference gauging station (type \code{character}).}
}
}
\usage{
df.flys_sections
}
\description{
This dataset relates the reference gauging stations to river
  stationing as used within FLYS3
}
\references{
\insertRef{bundesanstalt_fur_gewasserkunde_flys_2016}{hyd1d}
}
\keyword{datasets}
