% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterLevelDataFrame-methods.R
\name{setGaugingStationsMissing<-}
\alias{setGaugingStationsMissing<-}
\alias{setGaugingStationsMissing<-,WaterLevelDataFrame,character-method}
\alias{setGaugingStationsMissing<--method}
\title{Set a WaterLevelDataFrame's slot gauging_stations_missing}
\usage{
setGaugingStationsMissing(x) <- value

\S4method{setGaugingStationsMissing}{WaterLevelDataFrame,character}(x) <- value
}
\arguments{
\item{x}{an object of class \linkS4class{WaterLevelDataFrame}.}

\item{value}{a new value of class \code{character} for the
\code{gauging_stations_missing} slot.}
}
\value{
The function above sets a new \code{value} for the slot
  \code{gauging_stations_missing} and returns an object of class
  \linkS4class{WaterLevelDataFrame}.
}
\description{
A function to set the slot \code{gauging_stations_missing} of 
  an object of class \linkS4class{WaterLevelDataFrame}.
}
\examples{
wldf <- WaterLevelDataFrame(river   = "Elbe",
                            time    = as.POSIXct("2016-12-21"),
                            station = seq(257, 262, 0.1))
setGaugingStationsMissing(wldf) <- as.character("VOCKERODE")

}
\seealso{
\code{\link{getGaugingStationsMissing-method}}
}
