% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterLevelDataFrame-methods.R
\name{summary.WaterLevelDataFrame}
\alias{summary.WaterLevelDataFrame}
\title{WaterLevelDataFrame summary}
\usage{
\method{summary}{WaterLevelDataFrame}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{WaterLevelDataFrame} for which
a summary is desired.}

\item{\dots}{additional arguments to be passed to internally used functions.}
}
\value{
A list of summary statistics of the \linkS4class{WaterLevelDataFrame}
  and its slots.
}
\description{
Returns a list of descriptive statistics for an object of class 
  \linkS4class{WaterLevelDataFrame}.
}
\examples{
wldf <- WaterLevelDataFrame(river   = "Elbe",
                            time    = as.POSIXct("2016-12-21"),
                            station = seq(257, 262, 0.1))
wldf <- waterLevel(wldf)
summary(wldf)

}
\seealso{
\code{\link[base]{summary}}
}
